+++ CHANGELOG.md
@@ -5,8 +5,7 @@
 - Protocol-level tokens:
   - Simplified token transaction handling: Consolidated all token holder and
     governance operations for a single `TokenUpdateTransaction` type.
-  - Moved authorization from the scheduler to the token module and moved the
-    token governance account from `PLTConfiguration` to the token module state.
+  - Moved authorization from the scheduler to the token module.
 - Fix a bug in checking authorization for CreatePLT update.
 
 ## 9.0.4 (DevNet)